requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
pydantic==2.5.0
PyMuPDF==1.23.8
paddleocr==2.7.3
paddlepaddle==2.6.0
Pillow==10.1.0
google-generativeai==0.8.3
playwright==1.40.0

